// src/app/interfaces/values.ts
export interface values {
    name: string;
    h1: number;
    h2: number;
    h3: number;
    h4: number;
    h5: number;
    h6: number;
    h7: number;
    h8: number;
    h9: number;
    h10: number;
    h11: number;
    h12: number;
    h13: number;
    h14: number;
    h15: number;
    h16: number;
    h17: number;
    h18: number;
    In: number;
    Out: number;
    Tot: number;  
}